data_sources/extract_load/requirements.txt
psycopg2-binary
python-dotenv
google-cloud-storage
google-cloud-bigquery